Extract Pip runtime patches.
This moves the Pip tool into its own package and extracts the Pip runtime
patching code into a sibling resource that can be type-checked and linted.

Fixes #1581
