+++ tests/integration/test_issue_539.py
@@ -8,7 +8,7 @@ import subprocess
 import pytest
 
 from pex.common import temporary_dir
-from pex.pip import get_pip
+from pex.pip.tool import get_pip
 from pex.testing import IS_PYPY, run_pex_command
