+++ tox.ini
@@ -82,6 +82,7 @@ commands =
 deps =
     attrs==21.2.0  # This version should track the version in pex/vendor/__init__.py.
     packaging==20.9  # This version should track the version in pex/vendor/__init__.py.
+    pip==20.3.4  # This version should track the version in pex/vendor/__init__.py.
     setuptools==44.0.0  # This version should track the version in pex/vendor/__init__.py.
     mypy[python2]==0.931
     typing-extensions
